test: fix and move test
